# openapi/openapi.yaml
openapi: 3.0.0
info:
  description: "schema"
  version: "1.0.0"
  title: "integration test"
  contact:
    name: "integration test"
servers:
  - url: "http://localhost:80"
    description: local
tags:
  - name: "user"
    description: ユーザー
security:
  - Bearer: []
paths:
  /user:
    get:
      tags:
        - user
      operationId: FindUsers
      summary: 全ユーザー取得
      description: 全ユーザー取得
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                title: Users
                type: array
                items: { $ref: "#/components/schemas/User" }
        500:
          description: Internal Server Error
    post:
      tags:
        - user
      operationId: CreateUser
      summary: ユーザー作成
      description: ユーザー作成
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ID"
        500:
          description: Internal Server Error
  /user/{id}:
    get:
      tags:
        - user
      operationId: GetUserByID
      summary: 指定したIDのユーザー取得
      description: 指定したIDのユーザー取得
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        500:
          description: Internal Server Error
    put:
      tags:
        - user
      operationId: UpdateUser
      summary: 指定したIDのユーザー更新
      description: 指定したIDのユーザー更新
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error
    delete:
      tags:
        - user
      operationId: DeleteUser
      summary: 指定したIDのユーザー削除
      description: 指定したIDのユーザー削除
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error

components:
  schemas:
    ID:
      title: ID
      type: object
      properties:
        id:
          type: integer
          format: int64
      required:
        - id
    User:
      title: User
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - name
        - email
        - created_at